[TRY-SERVER] Test ICS emulator build with gcc-4,8 based toolchain

+++ testing/mochitest/runtests.py
@@ -108,6 +108,10 @@ class MessageLogger(object):
     VALID_ACTIONS = set(['suite_start', 'suite_end', 'test_start', 'test_end',
                          'test_status', 'log',
                          'buffering_on', 'buffering_off'])
+    TEST_PATH_PREFIXES = ['/tests/',
+                          'chrome://mochitests/content/browser/',
+                          'chrome://mochitests/content/chrome/']
+
 
     def __init__(self, logger, buffering=True):
         self.logger = logger
@@ -124,6 +128,16 @@ class MessageLogger(object):
         """True if the given object is a valid structured message (only does a superficial validation)"""
         return isinstance(obj, dict) and 'action' in obj and obj['action'] in MessageLogger.VALID_ACTIONS
 
+    def _fix_test_name(self, message):
+      """Normalize a logged test path to match the relative path from the sourcedir.
+      """
+      if 'test' in message:
+        test = message['test']
+        for prefix in MessageLogger.TEST_PATH_PREFIXES:
+          if test.startswith(prefix):
+            message['test'] = test[len(prefix):]
+            break
+
     def parse_line(self, line):
         """Takes a given line of input (structured or not) and returns a list of structured messages"""
         line = line.rstrip().decode("UTF-8", "replace")
@@ -138,6 +152,7 @@ class MessageLogger(object):
                     message = dict(action='log', level='info', message=fragment, unstructured=True)
             except ValueError:
                 message = dict(action='log', level='info', message=fragment, unstructured=True)
+            self._fix_test_name(message)
             messages.append(message)
 
         return messages
@@ -671,9 +686,9 @@ class MochitestUtilsMixin(object):
       paths.append(test)
 
     # Bug 883865 - add this functionality into manifestparser
-    with open(os.path.join(SCRIPT_DIR, 'tests.json'), 'w') as manifestFile:
+    with open(os.path.join(SCRIPT_DIR, options.testRunManifestFile), 'w') as manifestFile:
       manifestFile.write(json.dumps({'tests': paths}))
-    options.manifestFile = 'tests.json'
+    options.manifestFile = options.testRunManifestFile
 
     return self.buildTestURL(options)
 
@@ -692,7 +707,7 @@ class MochitestUtilsMixin(object):
       with open(options.pidFile + ".xpcshell.pid", 'w') as f:
         f.write("%s" % self.server._process.pid)
 
-  def startServers(self, options, debuggerInfo):
+  def startServers(self, options, debuggerInfo, ignoreSSLTunnelExts = False):
     # start servers and set ports
     # TODO: pass these values, don't set on `self`
     self.webServer = options.webServer
@@ -710,7 +725,7 @@ class MochitestUtilsMixin(object):
     self.startWebSocketServer(options, debuggerInfo)
 
     # start SSL pipe
-    self.sslTunnel = SSLTunnel(options, logger=self.log)
+    self.sslTunnel = SSLTunnel(options, logger=self.log, ignoreSSLTunnelExts = ignoreSSLTunnelExts)
     self.sslTunnel.buildConfig(self.locations)
     self.sslTunnel.start()
 
@@ -857,7 +872,7 @@ overlay chrome://browser/content/browser.xul chrome://mochikit/content/jetpack-a
     return extensions
 
 class SSLTunnel:
-  def __init__(self, options, logger):
+  def __init__(self, options, logger, ignoreSSLTunnelExts = False):
     self.log = logger
     self.process = None
     self.utilityPath = options.utilityPath
@@ -867,6 +882,7 @@ class SSLTunnel:
     self.httpPort = options.httpPort
     self.webServer = options.webServer
     self.webSocketPort = options.webSocketPort
+    self.useSSLTunnelExts = not ignoreSSLTunnelExts
 
     self.customCertRE = re.compile("^cert=(?P<nickname>[0-9a-zA-Z_ ]+)")
     self.clientAuthRE = re.compile("^clientauth=(?P<clientauth>[a-z]+)")
@@ -892,7 +908,7 @@ class SSLTunnel:
         config.write("redirhost:%s:%s:%s:%s\n" %
                      (loc.host, loc.port, self.sslPort, redirhost))
 
-      if option in ('ssl3', 'rc4'):
+      if self.useSSLTunnelExts and option in ('ssl3', 'rc4'):
         config.write("%s:%s:%s:%s\n" % (option, loc.host, loc.port, self.sslPort))
 
   def buildConfig(self, locations):
@@ -1159,7 +1175,8 @@ class Mochitest(MochitestUtilsMixin):
     if options.browserChrome and options.timeout:
       options.extraPrefs.append("testing.browserTestHarness.timeout=%d" % options.timeout)
     options.extraPrefs.append("browser.tabs.remote.autostart=%s" % ('true' if options.e10s else 'false'))
-    options.extraPrefs.append("browser.tabs.remote.sandbox=%s" % options.contentSandbox)
+    if options.strictContentSandbox:
+        options.extraPrefs.append("security.sandbox.windows.content.moreStrict=true")
 
     # get extensions to install
     extensions = self.getExtensionsToInstall(options)
@@ -1313,9 +1330,10 @@ class Mochitest(MochitestUtilsMixin):
 
   def cleanup(self, options):
     """ remove temporary files and profile """
-    if self.manifest is not None:
+    if hasattr(self, 'manifest') and self.manifest is not None:
       os.remove(self.manifest)
-    del self.profile
+    if hasattr(self, 'profile'):
+        del self.profile
     if options.pidFile != "":
       try:
         os.remove(options.pidFile)
@@ -1694,7 +1712,7 @@ class Mochitest(MochitestUtilsMixin):
         testsToRun = bisect.pre_test(options, testsToRun, status)
         # To inform that we are in the process of bisection, and to look for bleedthrough
         if options.bisectChunk != "default" and not bisection_log:
-            log.info("TEST-UNEXPECTED-FAIL | Bisection | Please ignore repeats and look for 'Bleedthrough' (if any) at the end of the failure list")
+            self.log.info("TEST-UNEXPECTED-FAIL | Bisection | Please ignore repeats and look for 'Bleedthrough' (if any) at the end of the failure list")
             bisection_log = 1
 
       result = self.doTests(options, onLaunch, testsToRun)
@@ -1783,9 +1801,11 @@ class Mochitest(MochitestUtilsMixin):
     # TODO: use mozrunner.local.debugger_arguments:
     # https://github.com/mozilla/mozbase/blob/master/mozrunner/mozrunner/local.py#L42
 
-    debuggerInfo = mozdebug.get_debugger_info(options.debugger,
-                                              options.debuggerArgs,
-                                              options.debuggerInteractive)
+    debuggerInfo = None
+    if options.debugger:
+        debuggerInfo = mozdebug.get_debugger_info(options.debugger,
+                                                  options.debuggerArgs,
+                                                  options.debuggerInteractive)
 
     if options.useTestMediaDevices:
       devices = findTestMediaDevices(self.log)
@@ -1820,8 +1840,8 @@ class Mochitest(MochitestUtilsMixin):
       testURL = self.buildTestPath(options, testsToFilter)
 
       # read the number of tests here, if we are not going to run any, terminate early
-      if os.path.exists(os.path.join(SCRIPT_DIR, 'tests.json')):
-        with open(os.path.join(SCRIPT_DIR, 'tests.json')) as fHandle:
+      if os.path.exists(os.path.join(SCRIPT_DIR, options.testRunManifestFile)):
+        with open(os.path.join(SCRIPT_DIR, options.testRunManifestFile)) as fHandle:
           tests = json.load(fHandle)
         count = 0
         for test in tests['tests']:
@@ -2117,7 +2137,7 @@ class Mochitest(MochitestUtilsMixin):
 
   def getTestManifest(self, options):
     if isinstance(options.manifestFile, TestManifest):
-        manifest = options.manifestFile
+      manifest = options.manifestFile
     elif options.manifestFile and os.path.isfile(options.manifestFile):
       manifestFileAbs = os.path.abspath(options.manifestFile)
       assert manifestFileAbs.startswith(SCRIPT_DIR)
@@ -2132,6 +2152,8 @@ class Mochitest(MochitestUtilsMixin):
 
       if os.path.exists(masterPath):
         manifest = TestManifest([masterPath], strict=False)
+      else:
+        self._log.warning('TestManifest masterPath %s does not exist' % masterPath)
 
     return manifest
